requirements.txt
Pillow>=9.0
reportlab>=3.6
